# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AGOGO API",
    "description": "API for AGOGO's APP."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/dietitian/login": {
      "get": {
        "tags": ["Dietitian API"],
        "summary": "Login for Dietitian",
        "description": "Authenticate and retrieve details for a dietitian based on email and password.",
        "parameters": [
          {
            "name": "dietitian_email",
            "in": "body",
            "required": true,
            "description": "Email of the dietitian.",
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_email": {
                  "type": "string"
                },
                "dietitian_pwd": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Dietitian"
            }
          },
          "400": {
            "description": "Email or password is missing or incorrect, or other bad request."
          }
        }
      }
    },
    "/dietitian/patients": {
      "get": {
        "tags": ["Dietitian API"],
        "summary": "Fetch dietitian's patients",
        "description": "Fetch all patients of a specific dietitian",
        "parameters": [
          {
            "name": "dietitianDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
                "Type": "array",
                "items":{ 
                "$ref": "#/definitions/Patient"
                }
          }
        }
      }
     }
    },
    "/dietitian/createPatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Create a patient",
        "description": "Create a new patient for a specific dietitian",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
               "$ref": "#/definitions/CreatePatient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
               "$ref": "#/definitions/Patient"
             }
          }
        }
      }
    },
    "/dietitian/deactivatePatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Deactivate a patient",
        "description": "Deactivate a specific patient",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
                "patient_ID": {
                  "type": "String"
                }
              }
          }
        }
      }
    },
    "/dietitian/activatePatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Activate a patient",
        "description": "Activate a specific patient",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
                 "patient_ID": {
                   "type": "String"
                 }
               }
          }
        }
      }
    },
    "/MealPrep/generateMealPlan": {
      "get": {
        "tags": ["Meal Plan Generation API"],
        "summary": "Generate meal plan using LSM",
        "description": "Generates a meal plan based on protein, carbs, and fat goals using Least Squares Method (LSM).",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_ID": {
                  "type": "string"
                },
                "protein_goal": {
                  "type": "number"
                },
                "carbs_goal": {
                  "type": "number"
                },
                "fat_goal": {
                  "type": "number"
                },
                "nbr_days": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/GenMealPrep"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/MealPrep/generateMealPlanFixedLunch": {
      "get": {
        "tags": ["Meal Plan Generation API"],
        "summary": "Generate meal plan with fixed lunch",
        "description": "Generates a meal plan with a fixed lunch ID.",
        "parameters": [
          {
            "name": "mealPlanFixedLunchDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_ID": {
                  "type": "string"
                },
                "protein_goal": {
                  "type": "number"
                },
                "carbs_goal": {
                  "type": "number"
                },
                "fat_goal": {
                  "type": "number"
                },
                "nbr_days": {
                  "type": "integer"
                },
                "fixed_lunch_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/GenMealPrep"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/MealPrep/insertMealPlanItem": {
      "post": {
        "tags": ["Meal Plan Item API"],
        "summary": "Adds a Meal Plan Item to the DB",
        "description": "Adds a Meal Plan Item to the DB",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
                "$ref": "#/definitions/MealPrepItem"
              }
          }
        ],
        "responses": {
          "200": {
            "description": "meal prep item added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }, 
    "/MealPrep/insertBulkMPItems": {
      "post": {
        "tags": ["Meal Plan Item API"],
        "summary": "Adds several Meal Plan Items to the DB in one call",
        "description": "Adds several Meal Plan Items to the DB in one call",
        "parameters": [
          {
            "name": "insert_bulk_MP_Items",
            "in": "body",
            "required": true,
            "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/MealPrepItem"
                }
              }
          }
        ],
        "responses": {
          "200": {
            "description": "meal prep items added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }, 
    "/MealPrep/getCombination": {
      "get": {
        "tags": ["Meal Plan Item API"],
        "summary": "gets a specific saved combination based on the patient, diet, and combination ID",
        "description": "gets a specific saved combination based on the patient, diet, and combination ID",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "diet_id": {
                  "type": "integer",
                  "format": "int64"
                },
                "patient_id": {
                  "type": "integer",
                  "format": "int64"
                },
                "combination_id": {
                  "type": "integer",
                  "format": "int64"
                }}
              }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Combination"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/diet/getDietCombinations": {
      "get": {
        "tags": ["Meal Plan Item API"],
        "summary": "gets a all saved  combinations based on the patient and diet",
        "description": "gets a all saved  combinations based on the patient and diet",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "diet_id": {
                  "type": "integer",
                  "format": "int64"
                },
                "patient_id": {
                  "type": "integer",
                  "format": "int64"
                }}
              }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
                "type": "array",
                "items": {
                "$ref": "#/definitions/Combination"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/Recipee/getRecipee": {
      "get": {
        "tags": ["Recipee API"],
        "summary": "Fetch Recipee details",
        "description": "Fetch details of a Recipee by its ID.",
        "parameters": [
          {
            "name": "recipee_id",
            "in": "body",
            "required": true,
            "description": "ID of the recipee to be fetched.",
            "schema": {
              "type": "object",
              "properties": {
                "ingredient_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Recipee"
            }
          },
          "400": {
            "description": "recipee ID is required or other bad request."
          },
          "404": {
            "description": "recipee not found."
          }
        }
      }
    },
   "/Ingredient/details": {
      "get": {
        "tags": ["Ingredient API"],
        "summary": "Fetch ingredient details",
        "description": "Fetch details of a specific ingredient by its ID.",
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "body",
            "required": true,
            "description": "ID of the ingredient to be fetched.",
            "schema": {
              "type": "object",
              "properties": {
                "ingredient_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          },
          "400": {
            "description": "Ingredient ID is required or other bad request."
          },
          "404": {
            "description": "Ingredient not found."
          }
        }
      }
    },
    "/Ingredient/addIngredient": {
      "post": {
        "tags": ["Ingredient API"],
        "summary": "Add a new ingredient",
        "description": "Add a new ingredient to the database.",
        "parameters": [
          {
            "name": "ingredient_data",
            "in": "body",
            "required": true,
            "description": "Data of the ingredient to be added.",
            "schema": {
              "$ref": "#/definitions/CreateIngredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredient added successfully."
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/Ingredient/updateIngredient": {
      "put": {
        "tags": ["Ingredient API"],
        "summary": "Update an ingredient",
        "description": "Update details of a specific ingredient by its ID.",
        "parameters": [
          {
            "name": "ingredient_data",
            "in": "body",
            "required": true,
            "description": "Updated data of the ingredient.",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredient updated successfully."
          },
          "400": {
            "description": "Ingredient ID is required or other bad request."
          },
          "404": {
            "description": "Ingredient not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }, 
   "/Diet/createDiet": {
      "post": {
        "tags": ["Diet API"],
        "summary": "Create a new diet",
        "description": "Add a new diet record for a patient.",
        "parameters": [
          {
            "name": "diet_data",
            "in": "body",
            "required": true,
            "description": "Data for the diet to be created.",
            "schema": {
              "$ref": "#/definitions/CreateDiet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet added successfully with ID: {diet_id}"
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/Diet/udpateDiet": {
      "put": {
        "tags": ["Diet API"],
        "summary": "Update an existing diet",
        "description": "Update a diet record for a patient using the diet ID.",
        "parameters": [
          {
            "name": "diet_data",
            "in": "body",
            "required": true,
            "description": "Updated data for the diet.",
            "schema": {
              "$ref": "#/definitions/Diet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet updated successfully with ID: {diet_id}."
          },
          "400": {
            "description": "Bad request or Diet ID is required."
          },
          "404": {
            "description": "Diet not found."
          }
        }
      }
    },
    "/Diet/deleteDiet": {
      "delete": {
        "tags": ["Diet API"],
        "summary": "Delete a diet",
        "description": "Remove a diet record using the diet ID.",
        "parameters": [
          {
            "name": "diet_id",
            "in": "body",
            "required": true,
            "description": "ID of the diet to be deleted.",
            "schema": {
              "type": "object",
              "properties": {
                "diet_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet deleted successfully with ID: {diet_id}."
          },
          "400": {
            "description": "Diet ID is required or bad request."
          },
          "404": {
            "description": "Diet not found."
          }
        }
      }
    },
    "/Diet/getDietHistory": {
      "get": {
        "tags": ["Diet API"],
        "summary": "Fetch diet history for a patient",
        "description": "Retrieve all diet records for a given patient ID.",
        "parameters": [
          {
            "name": "patient_id",
            "in": "body",
            "required": true,
            "description": "ID of the patient.",
            "schema": {
              "type": "object",
              "properties": {
                "patient_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Diet"
              }
            }
          },
          "400": {
            "description": "Patient ID is required or bad request."
          },
          "404": {
            "description": "No diets found for the given patient ID."
          }
        }
      }
    },
    "/Diet/getLastDiet": {
      "get": {
        "tags": ["Diet API"],
        "summary": "Fetch the last diet for a patient",
        "description": "Retrieve the most recent diet record for a given patient ID.",
        "parameters": [
          {
            "name": "patient_id",
            "in": "body",
            "required": true,
            "description": "ID of the patient.",
            "schema": {
              "type": "object",
              "properties": {
                "patient_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Diet"
            }
          },
          "400": {
            "description": "Patient ID is required or bad request."
          },
          "404": {
            "description": "No diets found for the given patient ID."
          }
        }
      }
    },
    "/patient/LastLifeStyle": {
      "get": {
        "tags": ["LifeStyle API"],
        "summary": "Fetch the latest lifestyle for a patient",
        "description": "Retrieve the most recent lifestyle record for a given patient ID.",
        "parameters": [
          {
            "name": "patient_ID",
            "in": "body",
            "required": true,
            "description": "ID of the patient.",
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LifeStyle"
            }
          },
          "400": {
            "description": "Patient ID is required or bad request."
          }
        }
      }
    },
    "/patient/LifeStyleHistory": {
      "get": {
        "tags": ["LifeStyle API"],
        "summary": "Fetch lifestyle history for a patient",
        "description": "Retrieve all lifestyle records for a given patient ID.",
        "parameters": [
          {
            "name": "patient_ID",
            "in": "body",
            "required": true,
            "description": "ID of the patient.",
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LifeStyle"
              }
            }
          },
          "400": {
            "description": "Patient ID is required or bad request."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Dietitian": {
      "type": "object",
      "properties": {
        "dietitianID": {
          "type": "string",
          "description": "ID of the Dietitian."
        },
        "Fname": {
          "type": "string",
          "description": "First name of the Dietitian."
        },
        "Lname": {
          "type": "string",
          "description": "Last name of the Dietitian."
        },
        "DOB": {
          "type": "string",
          "format": "date",
          "description": "Date of birth of the Dietitian."
        },
        "phone": {
          "type": "string",
          "description": "Phone number of the Dietitian."
        },
        "email": {
          "type": "string",
          "description": "Email address of the Dietitian."
        }
             }
    },
...
  "Combination":{
    "type": "object",
    "properties": {
      "breakfast": {
          "$ref": "#/definitions/Recipee"
      },
      "lunch": {
          "$ref": "#/definitions/Recipee"
      },
      "dinner": {
          "$ref": "#/definitions/Recipee"
      },
      "score": {
        "type": "number",
        "description": "score of the combination."
      }
  }
}  ,
...
  "GenMealPrep":{
    "type": "object",
    "properties": {
      "best_combinations": {
        "type": "array",
        "description": "Array of Breakfast - Lunch - Dinner Combinations.",
        "items":{
                "$ref": "#/definitions/Combination"
              }
            }
          }
        },
...
      }
}